画面上の全ての文言を多言語化 削除する写真が存在しない場合の例外処理を追加

+++ src/main/java/jp/yoshida/photoadmin/common/constant/KeyWordsConstants.java
@@ -32,7 +32,7 @@ public class KeyWordsConstants {
 
     public static final int MAX_FILE_NAME_LENGTH = 50;
 
-    public static final int MAX_FILE_SIZE_COEFFICIENT = 1;
+    public static final long MAX_FILE_SIZE_COEFFICIENT = 1;
 
     public static final int MAX_META_DATA_TEXT_LENGTH = 50;
